Move random to top level
