+++ docs/api.txt
@@ -108,7 +108,7 @@ Random
 ------------------------
 
 .. automodule:: stellargraph.utils.random
-  :members: get_seed, set_seed
+  :members: generate_seed, set_seed
 
 
 Datasets
